Add code to intermediate_medium elf.
Add just the last 8 "rounds" of assigments between dummy variables from
simple_small elf. The size of main in intermediate_medium (in my build)
was only 2468B compared to the size of 3300B of main in simple_small
(althought a bigger variety of operations in intermediate_medium).
After this addition, the size of main in intermediate_medium becomes
5074B.