Trim intermediates to match main size in simple.
Having the main sizes approximately similar should provide a better
comparison between the impact of the complexity of the starting template
program and outcomes (e.g. when comparing intermediate_medium and
simple_mediun for e.g. 100_mnist_digits).

NOTE: The sizes of mains in simple and intermediate programs may be
compiler depentend and may not match exacly on different systems.