Select parents using scores kept in the Evolver.
This is an intermediate step in the migration of keeping scores from
Programs to the Evolver (where they belong). Original functionality
of keeping scores in Programs is still present with some asserts
comparing the scores in Programs with those kept in the Evolver. The
next step will be to remove the score keeping from Programs completely.